Split out authorization into new library.
