+++ src/RService.IO.Abstractions/project.json
@@ -25,7 +25,8 @@
 
   "dependencies": {
     "Microsoft.AspNetCore.Http.Abstractions": "1.0.0",
-    "Microsoft.AspNetCore.Routing.Abstractions": "1.0.0"
+    "Microsoft.AspNetCore.Routing.Abstractions": "1.0.0",
+    "Microsoft.Extensions.DependencyInjection.Abstractions": "1.0.0"
   },
 
   "frameworks": {
