+++ src/RService.IO.Authorization/project.json
@@ -1,0 +1,37 @@
+﻿{
+    "title": "RService.IO.Authorization",
+    "version": "0.3.0",
+
+    "packOptions": {
+        "summary": "RService.IO authorization provider.",
+        "releaseNotes": "Initial alpha release",
+        "owners": [ "James Stumme" ],
+        "tags": [
+            "microservice",
+            "webservice",
+            "dotnet core",
+            "api",
+            "RESTful",
+            "RService.IO"
+        ],
+        "licenseUrl": "https://github.com/Stoom/RService.IO/blob/master/LICENSE",
+        "projectUrl": "https://rservice.io",
+        "requireLicenseAcceptance": false,
+        "repository": {
+            "type": "git",
+            "url": "https://github.com/Stoom/RService.IO"
+        }
+    },
+
+    "dependencies": {
+        "Microsoft.AspNetCore.Authorization": "1.0.0",
+        "Microsoft.Extensions.SecurityHelper.Sources": "1.0.0-rtm-21431",
+        "RService.IO.Abstractions": "0.3.*"
+    },
+
+    "frameworks": {
+        "netstandard1.6": {
+            "imports": ["dotnet5.6", "portable-net45+win8"]
+        }
+    }
+}
