+++ src/RService.IO/DependencyIngection/ServiceCollectionExtensions.cs
@@ -47,17 +47,6 @@ namespace RService.IO.DependencyIngection
                 services.AddTransient(serviceType);
             }
 
-
-            return services;
-        }
-
-        public static IServiceCollection AddRServiceIoAuthorization(
-            this IServiceCollection services)
-        {
-            services.AddOptions();
-
-            services.TryAddTransient<IAuthProvider, AuthProvider>();
-
             return services;
         }
     }
