+++ test/RService.IO.Authorization.Tests/project.json
@@ -1,0 +1,35 @@
+{
+  "version": "1.0.0",
+
+  "testRunner": "xunit",
+
+  "dependencies": {
+    "coveralls.io": "1.3.4",
+    "dotnet-test-xunit": "2.2.0-*",
+    "OpenCover": "4.6.519",
+    "RService.IO": "0.3.*",
+    "RService.IO.Authorization": "0.3.*",
+    "Microsoft.Extensions.Logging": "1.0.0-*",
+    "Microsoft.Extensions.Logging.Testing": "1.0.0-*",
+    "xunit": "2.2.0-*"
+  },
+
+  "frameworks": {
+    "netcoreapp1.0": {
+      "imports": [
+        "dotnet5.6",
+        "portable-net45+win8"
+      ],
+
+      "dependencies": {
+        "Microsoft.NETCore.App": {
+          "version": "1.0.0-*",
+          "type": "platform"
+        },
+        "moq.netcore": "4.4.0-*",
+        "System.Diagnostics.TraceSource": "4.0.0-*",
+        "FluentAssertions": "4.13.0"
+      }
+    }
+  }
+}
