+++ test/RService.IO.Tests/DependencyIngection/ServiceCollectionExtensionTests.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Reflection;
 using FluentAssertions;
 using Microsoft.AspNetCore.Builder;
@@ -221,35 +220,6 @@ namespace RService.IO.Tests.DependencyIngection
             globalExceptionFilter.Should().BeNull();
         }
 
-        [Fact]
-        public void AddRServiceIoAuthorization__AddsRServiceProviderForIAuthProvider()
-        {
-            var services = new ServiceCollection();
-
-            services.AddAuthorization()
-                .AddRServiceIoAuthorization();
-
-            var app = BuildApplicationBuilder(services);
-            var provider = app.ApplicationServices.GetService<IAuthProvider>();
-
-            provider.Should().NotBeNull().And.BeOfType<AuthProvider>();
-        }
-
-        [Fact]
-        public void AddRServiceIoAuthorization__UserImplementationForIAuthProviderTakesPrecedence()
-        {
-            var services = new ServiceCollection();
-
-            services.AddTransient<IAuthProvider, AuthorizationProvider>()
-                .AddAuthorization()
-                .AddRServiceIoAuthorization();
-
-            var app = BuildApplicationBuilder(services);
-            var provider = app.ApplicationServices.GetService<IAuthProvider>();
-
-            provider.Should().NotBeNull().And.BeOfType<AuthorizationProvider>();
-        }
-
         private static IApplicationBuilder BuildApplicationBuilder(IServiceCollection services)
         {
             var builder = new Mock<IApplicationBuilder>();
@@ -277,19 +247,6 @@ namespace RService.IO.Tests.DependencyIngection
             }
 
             public string DehydrateResponse(object resDto)
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        private class AuthorizationProvider : IAuthProvider
-        {
-            public Task<bool> IsAuthorizedAsync(HttpContext ctx, ServiceMetadata metadata)
-            {
-                throw new NotImplementedException();
-            }
-
-            public Task<bool> IsAuthorizedAsync(HttpContext ctx, IEnumerable<object> authorizationFilters)
             {
                 throw new NotImplementedException();
             }
