+++ test/Rservice.IO.Tests.Integration/MiddlewareTests.cs
@@ -13,8 +13,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Moq;
 using RService.IO;
-using RService.IO.Abstractions;
 using RService.IO.Abstractions.Providers;
+using RService.IO.Authorization.DependencyInjection;
 using RService.IO.Tests;
 using Xunit;
 using RService.IO.DependencyIngection;
